requirements.txt
Flask==2.1.2
opencv-python==4.5.5.64
numpy==1.22.0
Pillow==9.0.1
tensorflow==2.8.0
Flask==2.1.2
opencv-python==4.5.5.64
numpy==1.22.0
Pillow==9.0.1
tensorflow==2.8.0
